flask_signup/requirements.txt
Flask==3.1.1
python-dotenv==1.0.1
requests==2.31.0
flask-cors==4.0.0
gunicorn
psycopg2-binary
